Clean: 変換処理、テンプレートをLineBotモデルに移動
