+++ app/models/line_bot.rb
@@ -1,0 +1,31 @@
+class LineBot
+  include ActiveModel::Model
+
+  ## オタク用語→標準語に変換する。
+  def converting_otaku(text)
+    if OtakuWord.pluck(:word).any?(text)
+      OtakuWord.find_by(word: text).meaning
+    else
+      ## マッチする単語がなかった場合に返すテキスト
+      '該当のオタク用語が見つかりませんでした。'
+    end
+  end
+
+  ## 標準語→オタク用語に変換する。
+  def converting_standard(text)
+    if OtakuWord.pluck(:meaning).any?(text)
+      OtakuWord.find_by(meaning: text).word
+    else
+      ## マッチする単語がなかった場合に返すテキスト
+      '該当の標準語が見つかりませんでした'
+    end
+  end
+
+  ## リプライメッセージ(JSON)の雛形
+  def messages_template(content)
+    {
+      type: "text",
+      text: "#{content}"
+    }
+  end
+end
